Remove the debug driver
The debug driver is a software (Python) endpoint for the CRTP protocol
that can be used to develop against when there is no accessable
hardware.

It requires maintenance in order to stay up-to-date with the CRTP
protocol and the Crazyflie and the benefit we receive from it does not
outweigh the amount work needed to keep it relevant.