log: Fix check for log config size
We need to check if the size of the log elements are larger than 26
rather than 30, since some bytes are taken up by timestamp and block id.

This mirrors the check done on firmware side. Without this a user will
get hard to track down errors when having LogConfig sizes between 27 and
30.