Here follow multiple minor updates to this step-by-step guide that increase user friendliness.
In "Step 2: Take off function", section "Changing the height":
In the first code snippet 'mc.stop()' was not added. However, in the full script one can use to double check against 'mc.stop()' is added. As this function is called automatically when the motion commander instance is exited, I'd drop it for clarity.

In "Step 4: Logging while flying":
If one follows the snippets for a step-wise instruction, the code will not compile. This is because 'lg_stab' should actually be called 'logconf'.

In "Step 5: Combine logging and motion commander", section "Back and forth with limits":
In the first snippet a function 'move_box_limit()' is introduced and the text suggests to place it above a non-existing function 'move_in_box_limit()'. I updated it with 'move_linear_simple()' in order to be consistent with the full ready-to-copy example.
The second snippet will not compile, because start_forward() is a function of the motion commander instance.
Furthermore, the three variables 'body_x_cmd', 'body_y_cmd', and 'max_vel' are neither introduced nor needed in this section (only in the following one). Hence I'd drop them here for clarity.