Generate python documentation using pdoc3
This commit generates API reference documentation using pdoc3.

The program pdoc3 will traverse the cflib module and generate
documentation based on the template file text.mako that we drop in the
template-dir.

The `text.mako` file is doing all the heavy lifting to create the
markdown we want.

For more information about pdoc3 and the custom template system see:
   https://pdoc3.github.io/pdoc/doc/pdoc/#custom-templates