wireshark: Add toc handling to parameters
This adds dissection of the toc channel of the parameter port which
enables us to know the group and name for parameter reads and writes.

You can now filter on the name of a group or a name and follow it
through your session. And also check the type of the parameter.