Increase ESP flashing timeout by 5s to match GAP8.
When flashing the ESP, we also reboot the GAP8, requiring the same delay. However, since flashing the ESP firmware does not necessarily mean we are flashing the GAP8, we must explicitly check if the ESP is among our targets to apply the timeout correctly. This ensures the GAP8 is never in an incorrect state.

This fixes #526.

Also adds a missing boot delay in the reset-to-firmware call, originally intended in b7ead1937609995ea6ab7f9a3d87fd2d269eac73.