Close the param threads properly if close is called.
Put a None object on the queue if closed and read this out.
The lock was released but since the queue was emptied the run task would never return
from queue.get() and then this thread could not be closed. Make sure to
unblock it by sending None and then setting should_close variable